requirements.txt
amqp==5.2.0
asgiref==3.7.2
astroid==3.0.1
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
dill==0.3.7
Django==4.2.7
djangorestframework==3.14.0
gunicorn==21.2.0
isort==5.12.0
kombu==5.3.4
mccabe==0.7.0
mysqlclient==2.2.0
packaging==23.2
platformdirs==4.1.0
prompt-toolkit==3.0.41
pycodestyle==2.11.1
PyMySQL==1.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.1
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
tomlkit==0.12.3
typing_extensions==4.8.0
tzdata==2023.3
vine==5.1.0
wcwidth==0.2.12
